requirements.txt
requests>=2.31.0
paho-mqtt>=1.6.1
sqlalchemy>=2.0.0
pymysql>=1.1.0
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
scikit-learn>=1.3.0
prophet>=1.1.4
fastapi>=0.100.0
uvicorn>=0.23.0
jupyter>=1.0.0
pytest>=7.4.0
paho-mqtt
cryptography